:recycle: Combine dev token tests
